flask/requirements.txt
flask
mysql-connector-python
python-dotenv
werkzeug
pytz